+++ tests/open_alchemy/test_build.py
@@ -3,6 +3,7 @@ import pytest
 
 from open_alchemy import build
 from open_alchemy import exceptions
+from open_alchemy.build import PackageFormat
 
 
 @pytest.mark.parametrize(
@@ -491,10 +492,27 @@ def test_calculate_version(spec, spec_str, expected_version):
     assert returned_version == expected_version
 
 
+@pytest.mark.parametrize(
+    "package_format, extensions",
+    [
+        pytest.param(PackageFormat.NONE, [], id="build without package"),
+        pytest.param(PackageFormat.SDIST, [".tar.gz"], id="build with a sdist package"),
+        pytest.param(
+            PackageFormat.WHEEL,
+            [".whl"],
+            id="build with a wheel package",
+        ),
+        pytest.param(
+            PackageFormat.SDIST | PackageFormat.WHEEL,
+            [".tar.gz", "whl"],
+            id="build with all available packages",
+        ),
+    ],
+)
 @pytest.mark.build
-def test_execute(tmp_path):
+def test_execute(tmp_path, package_format, extensions):
     """
-    GIVEN spec, name and path
+    GIVEN spec, name, path and a package format
     WHEN execute is called with the spec, name and path
     THEN the setup.py, MANIFEST.in, spec.json and __init__.py files are created.
     """
@@ -518,7 +536,7 @@ def test_execute(tmp_path):
         },
     }
 
-    build.execute(spec=spec, name=name, path=str(dist))
+    build.execute(spec=spec, name=name, path=str(dist), format_=package_format)
 
     # Define generated project directories
     project_path = dist / name
@@ -585,3 +603,69 @@ init_json(parent_path / "spec.json")"""
     assert "class TSchema" in init_contents
     assert "id: 'sqlalchemy.Column[typing.Optional[int]]'" in init_contents
     assert "Schema: typing.Type[TSchema]" in init_contents
+
+    # Assert one package per requested format is created.
+    for extension in extensions:
+        dist_dir = project_path / "dist"
+        files = list(dist_dir.glob(f"{name}*{extension}"))
+        assert len(files) == 1
+
+
+@pytest.mark.integration
+def test_build_dist_wheel_import_error(tmp_path):
+    """
+    GIVEN wheel package is not available
+    WHEN build_dist is called
+    THEN RuntimeError is raised.
+    """
+    dist = tmp_path / "dist"
+    dist.mkdir()
+
+    name = "app_models"
+    version = "version 1"
+    spec = {
+        "info": {
+            "version": version,
+        },
+        "components": {
+            "schemas": {
+                "Schema": {
+                    "type": "object",
+                    "x-tablename": "schema",
+                    "properties": {"id": {"type": "integer"}},
+                }
+            }
+        },
+    }
+
+    try:
+        build.run("pip uninstall -y wheel", ".")
+        with pytest.raises(RuntimeError):
+            build.execute(
+                spec=spec, name=name, path=str(dist), format_=PackageFormat.WHEEL
+            )
+    finally:
+        build.run("pip install wheel", ".")
+
+
+@pytest.mark.build
+def test_validate_dist_format_valid():
+    """
+    GIVEN a valid format
+    WHEN validate_dist_format is called
+    THEN nothing happens
+    """
+    for format_ in PackageFormat:
+        build.validate_dist_format(format_)
+
+
+@pytest.mark.build
+def test_validate_dist_format_invalid():
+    """
+    GIVEN a invalid format
+    WHEN validate_dist_format is called
+    THEN an exception is raised
+    """
+    with pytest.raises(exceptions.BuildError):
+        format_ = PackageFormat.NONE | PackageFormat.SDIST
+        build.validate_dist_format(format_)